enhance msg
